Change several GetDropOperation methods return type
This CL changes the return type of MenuDelegate::GetDropOperation and
related methods from integer to DragOperation, to improve type safety.

Bug: 1093536
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2682441
Commit-Queue: Henrique Ferreiro <hferreiro@igalia.com>
Cr-Commit-Position: refs/heads/master@{#853652}

+++ ui/views/controls/menu/menu_controller.cc
@@ -1060,8 +1060,9 @@ int MenuController::OnDragUpdated(SubmenuView* source,
       query_menu_item = menu_item->GetParentMenuItem();
       drop_position = MenuDelegate::DropPosition::kOn;
     }
-    drop_operation = menu_item->GetDelegate()->GetDropOperation(
-        query_menu_item, event, &drop_position);
+    drop_operation =
+        static_cast<int>(menu_item->GetDelegate()->GetDropOperation(
+            query_menu_item, event, &drop_position));
 
     // If the menu has a submenu, schedule the submenu to open.
     SetSelection(menu_item, menu_item->HasSubmenu() ? SELECTION_OPEN_SUBMENU
